+++ Snoop.Core/Infrastructure/Helpers/ResourceDictionaryKeyHelpers.cs
@@ -8,38 +8,46 @@ namespace Snoop.Infrastructure.Helpers
     using System.Linq;
     using System.Windows;
     using System.Windows.Media;
-    using System.Windows.Media.Media3D;
 
     public static class ResourceDictionaryKeyHelpers
     {
-        public static string GetKeyOfResourceItem(DependencyObject? dependencyObject, object? resourceItem)
+        public static object GetKeyOfResourceItem(DependencyObject? dependencyObject, object? resourceItem)
         {
             if (dependencyObject is null
                 || resourceItem is null)
             {
-                return string.Empty;
+                return DependencyProperty.UnsetValue;
             }
 
             // Walk up the visual tree, looking for the resourceItem in each frameworkElement's resource dictionary.
-            while (dependencyObject is Visual or Visual3D)
+            while (dependencyObject is not null)
             {
-                if (dependencyObject is FrameworkElement frameworkElement)
+                if (dependencyObject is FrameworkElement fe)
                 {
-                    var resourceKey = GetKeyInResourceDictionary(frameworkElement.Resources, resourceItem);
+                    var resourceKey = GetKeyInResourceDictionary(fe.Resources, resourceItem)
+                                      ?? GetKeyInResourceDictionary(FrameworkElementHelper.GetStyle(fe)?.Resources, resourceItem)
+                                      ?? GetKeyInResourceDictionary(FrameworkElementHelper.GetTemplate(fe)?.Resources, resourceItem)
+                                      ?? GetKeyInResourceDictionary(FrameworkElementHelper.GetThemeStyle(fe)?.Resources, resourceItem);
                     if (resourceKey is not null)
                     {
                         return resourceKey;
                     }
                 }
-                else
+                else if (dependencyObject is FrameworkContentElement fce)
                 {
-                    break;
+                    var resourceKey = GetKeyInResourceDictionary(fce.Resources, resourceItem)
+                                      ?? GetKeyInResourceDictionary(FrameworkElementHelper.GetStyle(fce)?.Resources, resourceItem)
+                                      ?? GetKeyInResourceDictionary(FrameworkElementHelper.GetThemeStyle(fce)?.Resources, resourceItem);
+                    if (resourceKey is not null)
+                    {
+                        return resourceKey;
+                    }
                 }
 
-                dependencyObject = VisualTreeHelper.GetParent(dependencyObject);
+                dependencyObject = VisualTreeHelper.GetParent(dependencyObject) ?? LogicalTreeHelper.GetParent(dependencyObject);
             }
 
-            // check the application resources
+            // Check application resources
             if (Application.Current is not null)
             {
                 var resourceKey = GetKeyInResourceDictionary(Application.Current.Resources, resourceItem);
@@ -49,26 +57,51 @@ namespace Snoop.Infrastructure.Helpers
                 }
             }
 
-            return string.Empty;
+            // Check system resources
+            foreach (var cacheEntry in SystemResourcesCache.Instance.SystemResources.Reverse())
+            {
+                {
+                    var resourceKey = GetKeyInResourceDictionary(cacheEntry.Themed, resourceItem);
+                    if (resourceKey is not null)
+                    {
+                        return resourceKey;
+                    }
+                }
+
+                {
+                    var resourceKey = GetKeyInResourceDictionary(cacheEntry.Generic, resourceItem);
+                    if (resourceKey is not null)
+                    {
+                        return resourceKey;
+                    }
+                }
+            }
+
+            return DependencyProperty.UnsetValue;
         }
 
-        public static string? GetKeyInResourceDictionary(ResourceDictionary dictionary, object? resourceItem)
+        public static object? GetKeyInResourceDictionary(ResourceDictionary? dictionary, object? resourceItem)
         {
+            if (dictionary is null)
+            {
+                return null;
+            }
+
             foreach (var key in dictionary.Keys)
             {
                 if (dictionary.TryGetValue(key, out var item)
                     && item == resourceItem)
                 {
-                    return key?.ToString();
+                    return key;
                 }
             }
 
-            foreach (var dic in dictionary.MergedDictionaries.ToList())
+            foreach (var dic in dictionary.MergedDictionaries.Reverse())
             {
-                var name = GetKeyInResourceDictionary(dic, resourceItem);
-                if (!string.IsNullOrEmpty(name))
+                var key = GetKeyInResourceDictionary(dic, resourceItem);
+                if (key is not null)
                 {
-                    return name;
+                    return key;
                 }
             }